% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{div}
\alias{rbdd_prop_div}
\alias{feat_prop_div}
\alias{tuol_prop_div}
\alias{amer_prop_div}
\alias{moke_prop_div}
\alias{stan_prop_div}
\alias{rbdd_tot_div}
\alias{feat_tot_div}
\alias{tuol_tot_div}
\alias{amer_tot_div}
\alias{moke_tot_div}
\alias{stan_tot_div}
\title{Proportion and Total Diverted}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 216 rows and 3 columns.}
\usage{
rbdd_prop_div

feat_prop_div

tuol_prop_div

amer_prop_div

moke_prop_div

stan_prop_div

rbdd_tot_div

feat_tot_div

tuol_tot_div

amer_tot_div

moke_tot_div

stan_tot_div
}
\description{
Monthly propotion and total diverted values for the duration of screw trap observations.
}
\details{
Available screw trap data start and end years:
\tabular{lll}{
  \strong{Location} \tab \strong{start} \tab \strong{end}\cr
  RBDD \tab 1999  \tab 2016\cr
  Feather River\tab	1998 \tab	2016\cr
  Tuolumne River\tab	2007 \tab	2017\cr
  American River\tab	2013 \tab	2016\cr
  Battle Creek\tab	1998 \tab	2016\cr
  Clear Creek\tab	1998 \tab	2016\cr
  Mokelumne River\tab	1999 \tab	2015\cr
  Stanislaus River \tab	1998 \tab	2016
}
}
\section{Proportion Diverted}{

Values pre September 2003 are CALSIM model output and values after are the median
CALSIM value grouped by water year type and month.

For more details on CALSIM see:
\itemize{
  \item use this link within R \code{\link[cvpiaFlow]{proportion_diverted}}
  \item use this \href{https://flowwest.github.io/cvpiaFlow/reference/proportion_diverted.html}{link} if in a web browser
}

NOTE: RBDD is represented using the Upper Sacramento River DSM model node.
There are no modeled diversions for Battle and Clear Creek.
}

\section{Total Diverted}{

Values were estimated by multiplying the measured flow data and the modeled
proportion diverted values.
}

\keyword{datasets}
