% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flow_temperature}
\alias{feather_flow}
\alias{amer_flow}
\alias{moke_flow}
\alias{stan_flow}
\alias{clear_flow}
\alias{feather_temp}
\alias{amer_temp}
\alias{moke_temp}
\alias{stan_temp}
\alias{clear_temp}
\alias{feather_daily_flow}
\alias{amer_daily_flow}
\alias{moke_daily_flow}
\alias{stan_daily_flow}
\alias{clear_daily_flow}
\alias{feather_daily_temp}
\alias{amer_daily_temp}
\alias{moke_daily_temp}
\alias{stan_daily_temp}
\alias{clear_daily_temp}
\title{Gaged Flow and Temperature}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 241 rows and 3 columns.}
\usage{
feather_flow

amer_flow

moke_flow

stan_flow

clear_flow

feather_temp

amer_temp

moke_temp

stan_temp

clear_temp

feather_daily_flow

amer_daily_flow

moke_daily_flow

stan_daily_flow

clear_daily_flow

feather_daily_temp

amer_daily_temp

moke_daily_temp

stan_daily_temp

clear_daily_temp
}
\description{
Measured daily and monthly mean flow (cfs) and temperature (°C) from USGS and CDEC gages
for the duration of screw trap observations.
}
\details{
Available screw trap data start and end years:
\tabular{ll}{
  \strong{Location} \tab \strong{start-end}\cr
  Feather River\tab	1998 - 2016\cr
  American River\tab	2013 - 2016\cr
  Clear Creek\tab	1998 - 2016\cr
  Mokelumne River\tab	1999 - 2015\cr
  Stanislaus River \tab	1998 - 2016
}
}
\section{Flow and Temperature Gage IDs}{


\strong{Flow Gages:}
\itemize{
  \item Feather River - CDEC FEATHER RIVER NEAR GRIDLEY (GRL)
  \item American River - USGS AMERICAN R A FAIR OAKS CA (11446500)
  \item Clear Creek - CDEC CLEAR CREEK NEAR IGO (IGO)
  \item Mokelumne River - USGS MOKELUMNE R BL CAMANCHE DAM CA (11323500)
  \item Stanislaus River - USGS STANISLAUS R A RIPON CA (11303000)
}

\strong{Temperature Gages:}
\itemize{
  \item Feather River - CDEC FEATHER RIVER NEAR GRIDLEY (GRL)
  \item American River - USGS AMERICAN R A FAIR OAKS CA (11446500)
  \item Clear Creek  - CDEC CLEAR CREEK NEAR IGO (IGO)
  \item Mokelumne River - EDMUD provided data near Victor CA
  \item Stanislaus River - USGS STANISLAUS R A RIPON CA (11303000)
}

\strong{NOTE:} Feather temperature data available 2003 - 2006
For the American and Stanislaus Rivers the daily mean temperature
was represented using the mean of the maximum and minimum daily temperature.
}

\section{Daily Data}{

Daily data for flow and temperature are also available: \code{[watershed]_daily_flow}
or  \code{[watershed]_daily_temp}
}

\keyword{datasets}
