% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inputs}
\alias{fall_inputs}
\alias{winter_inputs}
\alias{spring_inputs}
\alias{steelhead_inputs}
\title{Synthetic Years Model inputs}
\format{list}
\usage{
fall_inputs

winter_inputs

spring_inputs

steelhead_inputs
}
\description{
All the model inputs for sythetic years to cover 1998-2017
}
\details{
Available screw trap data start and end years:

\tabular{ll}{
  \strong{Location} \tab \strong{start-end}\cr
  Feather River\tab	1998 - 2016\cr
  American River\tab	2013 - 2016\cr
  Clear Creek\tab	1998 - 2016\cr
  Mokelumne River\tab	1999 - 2015\cr
  Stanislaus River \tab	1998 - 2016
}


\strong{NOTE:}
\enumerate{
  \item Habitat values were scaled using the proportion of spawning and rearing extent above the trap.
  \item Apply a suitability factor to floodplain areas (suggested value = .27).
}
}
\keyword{datasets}
