% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flow_temperature}
\alias{rbdd_flow}
\alias{feather_flow}
\alias{tuol_flow}
\alias{amer_flow}
\alias{moke_flow}
\alias{stan_flow}
\alias{battle_flow}
\alias{clear_flow}
\alias{deer_flow}
\alias{rbdd_temp}
\alias{feather_temp}
\alias{tuol_temp}
\alias{amer_temp}
\alias{moke_temp}
\alias{stan_temp}
\alias{battle_temp}
\alias{clear_temp}
\alias{deer_temp}
\title{Monthly Mean Flow and Temperature}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 216 rows and 3 columns.}
\usage{
rbdd_flow

feather_flow

tuol_flow

amer_flow

moke_flow

stan_flow

battle_flow

clear_flow

deer_flow

rbdd_temp

feather_temp

tuol_temp

amer_temp

moke_temp

stan_temp

battle_temp

clear_temp

deer_temp
}
\description{
Measured monthly mean flow (cfs) and temperature (°C) from USGS and CDEC gages
for the duration of screw trap observations.
}
\details{
Available screw trap data start and end years:
\tabular{lll}{
  \strong{Location} \tab \strong{start} \tab \strong{end}\cr
  RBDD \tab 1999  \tab 2016\cr
  Feather River\tab	1998 \tab	2016\cr
  Tuolumne River\tab	2007 \tab	2017\cr
  American River\tab	2013 \tab	2016\cr
  Battle Creek\tab	1998 \tab	2016\cr
  Clear Creek\tab	1998 \tab	2016\cr
  Deer Creek\tab 1992 \tab 2010\cr
  Mokelumne River\tab	1999 \tab	2015\cr
  Stanislaus River \tab	1998 \tab	2016
}
}
\section{Flow and Temperature Gage IDs}{


\strong{Flow Gages:}
\itemize{
  \item RBDD (Upper Sacramento River) - USGS SACRAMENTO R AB BEND BRIDGE NR RED BLUFF CA (11377100)
  \item Feather River - CDEC FEATHER RIVER NEAR GRIDLEY (GRL)
  \item Tuolumne River - USGS TUOLUMNE R A MODESTO CA (11290000)
  \item American River - USGS AMERICAN R A FAIR OAKS CA (11446500)
  \item Battle Creek - USGS BATTLE C BL COLEMAN FISH HATCHERY NR COTTONWOOD CA (11376550)
  \item Clear Creek - CDEC CLEAR CREEK NEAR IGO (IGO)
  \item Deer Creek - USGS DEER C NR VINA CA (11383500)
  \item Mokelumne River - USGS MOKELUMNE R BL CAMANCHE DAM CA (11323500)
  \item Stanislaus River - USGS STANISLAUS R A RIPON CA (11303000)
}

\strong{Temperature Gages:}
\itemize{
  \item RBDD (Upper Sacramento River) - CDEC SACRAMENTO R AT RED BLUFF DIVERSION DAM (RDB)
  \item Feather River - CDEC FEATHER RIVER NEAR GRIDLEY (GRL)
  \item Tuolumne River - USGS TUOLUMNE R A MODESTO CA (11290000)
  \item American River - USGS AMERICAN R A FAIR OAKS CA (11446500)
  \item Battle Creek - USGS BATTLE C BL COLEMAN FISH HATCHERY NR COTTONWOOD CA (11376550)
  \item Clear Creek  - CDEC CLEAR CREEK NEAR IGO (IGO)
  \item Deer Creek - USGS DEER C NR VINA CA (11383500)
  \item Mokelumne River - EDMUD provided data near Victor CA
  \item Stanislaus River - USGS STANISLAUS R A RIPON CA (11303000)
}

\strong(NOTE:) Deer Creek temperature data begins in 1998
}

\keyword{datasets}
